Add partitioning to the `event` table.
This requires including the `time` column in `event`'s PK.
All of the stored procedure have been changed to pass around a `p_now`
parameter to ensure that the time value is the same in each
append (and `open_store`).

No `timestamp` columns should use a default value any longer.